+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 x.y.z Release notes (yyyy-MM-dd)
 =============================================================
 ### Enhancements
+* Add Xcode 16.4 binaries.
 * Add support for building with Xcode 26 beta 1.
 
 ### Fixed
@@ -12,7 +13,7 @@ x.y.z Release notes (yyyy-MM-dd)
 ### Compatibility
 * Realm Studio: 15.0.0 or later.
 * APIs are backwards compatible with all previous releases in the 10.x.y series.
-* Carthage release for Swift is built with Xcode 16.3.0.
+* Carthage release for Swift is built with Xcode 16.4.0.
 * CocoaPods: 1.10 or later.
 * Xcode: 15.3.0-26 beta 1.